+++ CHANGES.rst
@@ -8,8 +8,9 @@ Changelog
 - Changed from 90° to 270° image rotation in `BaseDGHV.image_orientation` because it is
   rotated clockwise with imagemagick, in pod templates including annexes.
   [aduchene]
-- Manage `MeetingConfig.defaultAdviceHiddenDuringRedaction` only when a new advice is added,
-  not when advice is asked_again, see https://support.imio.be/browse/PM-3883.
+- Manage `MeetingConfig.defaultAdviceHiddenDuringRedaction` when a new advice is added,
+  and when advice is asked_again the same way (in the edit form) and display a message
+  to the adviser.
   [gbastien]
 - Display `global_actions` on the advice view.
   [gbastien]